LayoutTests/http/tests/iframe-memory-monitor/media-video-autoplay-in-frames-expected.txt

PASS document.querySelector('iframe[name=frame1]').srcdoc is not ""
PASS successfullyParsed is true

TEST COMPLETE
